Modify /jobinfo, modify language bundle.
This commit modifies the /jobinfo command to correctly display the new information. It also modifies the command to make listed blocks appear more consistent in the list. A colon : was added at the end of each prefix, "Mine:, Fish:, Kill, Breed:, to provide a cleaner look for the command. en_US language bundle is modified, other languages are not yet changed.