Fix build.
Remove source files duplicating com.mojang.authlib because why were they there to begin with? These files are provided by Spigot as compiled code. Mark spigot artifact as 'provided' instead of 'test' in pom to fix the build. Previously the plugin was failing to load with a NoSuchMethodException because of compiler confusion over these conflicting files.