Bugfix for issue #98
This is a small fix to the system that load player banks. Previously if for whatever reason a player name existed in PlayerFile.yml that had neither a legacy bank account nor a modern iban it would retrieve a null iban value and then attempt to throw a GeneralEconomyException passed with a null value thus causing the plugin to crash with an NPE instead of handling the exception. New code checks for a null iban state and creates a fresh iban account with 0 currency.