Another bugfix.
Commit bugfix for NPE crash when loading player shops.

ItemStack stack = getShopInventory().getItem(slot).clone()

This method is called and iterates on every item slot in the player shop inventory. When a stack contains nothing it is technically null. Instantiating the local variable in this way causes the code to pass a null object to the .clone() function which crashes with an unchecked NPE instead of just instantiating the stack as null.

Fix is to check for itemstack != null and then clone the stack.